+++ dialogy/base/input/__init__.py
@@ -78,12 +78,13 @@ If there is a need to represent an :ref:`Input<Input>` as a `dict` we can do the
 """
 from __future__ import annotations
 
-from typing import Any, Dict, List, Optional, Set
+from functools import reduce
+from typing import Any, Dict, List, Optional, Union, Set
 
 import attr
 
 from dialogy.types import Utterance
-from dialogy.utils import is_unix_ts, normalize
+from dialogy.utils import is_unix_ts, normalize, get_best_transcript
 
 
 @attr.frozen
@@ -119,6 +120,11 @@ class Input:
     """
     A derived attribute. We cross product each utterance and return a list of strings.
     We use this to :ref:`normalize <normalize>` utterances.
+    """
+
+    best_transcript: str = attr.ib(default=None)
+    """
+    A derived attribute. Contains the best alternative selected out of the utterances.
     """
 
     clf_feature: Optional[List[str]] = attr.ib(  # type: ignore
@@ -212,6 +218,15 @@ class Input:
     Points at the active state (or node) within the conversation graph.
     """
 
+    nls_label: Optional[str] = attr.ib(
+        default=None,
+        kw_only=True,
+        validator=attr.validators.optional(attr.validators.instance_of(str)),
+    )
+    """
+    Points at the NLS Label of the active (current) node within the conversation graph.
+    """
+
     expected_slots: Set[str] = attr.ib(factory=set, kw_only=True)
     """
     In a given turn, the expected number of slots to fill.
@@ -230,7 +245,17 @@ class Input:
     def __attrs_post_init__(self) -> None:
         try:
             object.__setattr__(self, "transcripts", normalize(self.utterances))
-            object.__setattr__(self, "expected_slots", set(self.expected_slots) if self.expected_slots else None)
+
+            object.__setattr__(
+                self, "best_transcript", get_best_transcript(self.transcripts)
+            )
+
+            object.__setattr__(
+                self,
+                "expected_slots",
+                set(self.expected_slots) if self.expected_slots else None,
+            )
+
         except TypeError:
             ...
 
@@ -271,3 +296,36 @@ class Input:
         if reference:
             return attr.evolve(reference, **d)
         return attr.evolve(cls(utterances=d["utterances"]), **d)  # type: ignore
+
+    def find_entities_in_history(
+        self,
+        intent_names: Optional[List[str]] = None,
+        slot_names: Optional[List[str]] = None,
+    ) -> Union[List[Dict[str, Any]], None]:
+        if not self.history:
+            return None
+
+        _intent_names: List[str] = intent_names or []
+        _slot_names: List[str] = slot_names or []
+
+        # Flatten the history to a list of intents
+        intents: List[Dict[str, Any]] = reduce(
+            lambda intents, res: intents + res["intents"], self.history[::-1], []
+        )
+
+        # Filter intents by name
+        required_intents = filter(
+            lambda intent: intent["name"] in _intent_names, intents
+        )
+
+        # Flatten the intents to a list of slots
+        slots: List[Dict[str, Any]] = reduce(
+            lambda slots, intent: slots + intent["slots"], required_intents, []
+        )
+
+        # Filter slots by name
+        required_slots = filter(lambda slot: slot["name"] in _slot_names, slots)
+
+        return list(
+            reduce(lambda entities, slot: entities + slot["values"], required_slots, [])
+        )