+++ dialogy/constants/__init__.py
@@ -41,6 +41,8 @@ NUMBER = "number"
 AMOUNT_OF_MONEY = "amount-of-money"
 DURATION = "duration"
 CREDIT_CARD_NUMBER = "credit-card-number"
+ADDRESS = "address"
+PINCODE = "pincode"
 
 FUTURE = "future"
 PAST = "past"
@@ -142,6 +144,8 @@ ID = "id"
 IDX = "idx"
 DATA = "data"
 STATE = "state"
+LANG = "lang"
+NLS_LABEL = "nls_label"
 OUTPUT = "output"
 INPUT = "input"
 CLASSIFICATION_INPUT = "classification_input"
@@ -159,6 +163,7 @@ SCORE = "score"
 LABELENCODER_FILE = "labelencoder.pkl"
 INVALID_TOKENS = ["<UNK>", "<PAD>", "<SOS>", "<EOS>"]
 OUTPUT_ATTRIBUTES = {INTENTS, ENTITIES, ORIGINAL_INTENT, RUNTIMES}
+NULL_PROMPT_TOKEN = "<pad>"
 
 # MLP
 MLPMODEL_FILE = "mlpmodelpipeline.joblib"